Renamed the raw labour_state variable as labour_state_raw, and added a function to generate_composite_vars.py to generate labour_state according to the SIPHER 7 employment categories. ALSO refactored housing_quality from numeric to string factors (high, medium, low) as a bit of a test. Too confusing in numbers

+++ minos/data_generation/generate_stock_pop.py
@@ -159,7 +159,7 @@ def generate_stock(projections, cross_validation):
     data['ncigs'] = data['ncigs'].astype('int64')
     data['neighbourhood_safety'] = data['neighbourhood_safety'].astype('int64')
     data['nutrition_quality'] = data['nutrition_quality'].astype('int64')
-    data['housing_quality'] = data['housing_quality'].astype('int64')
+    #data['housing_quality'] = data['housing_quality'].astype('int64')
 
     US_utils.save_multiple_files(data, years, "data/final_US/", "")
